requirements.txt
django
psycopg2
djangorestframework
django-filter
django-cors-headers
dj-rest-auth
drf-yasg